requirements.txt
Django~=5.2.7
djangorestframework~=3.16.1
psycopg2-binary~=2.9.11
drf-spectacular~=0.29.0
python-decouple~=3.8
gunicorn~=23.0.0
python-dateutil~=2.9.0.post0
whitenoise~=6.11.0